refactor: 메세지 출력 함수 위치 변경
