+++ client/src/components/Header/Header.tsx
@@ -3,6 +3,9 @@ import { Box, Button } from "@mui/material";
 import { Link, useMatch } from "react-router-dom";
 import { S } from "./styled";
 import { Helmet } from "react-helmet-async";
+import useToastMessage from "@/hooks/common/useToastMessage";
+import { TOAST_MESSAGE } from "@/utils/toastMessage";
+import React from "react";
 
 const Header = () => {
   const { token: authToken, removeToken } = useAuthStore();
@@ -11,6 +14,7 @@ const Header = () => {
   const onLogout = () => {
     if (authToken) {
       removeToken("token");
+      useToastMessage(TOAST_MESSAGE.AUTH.LOGOUT_SUCCESS, "error");
     }
   };
   return (
@@ -43,4 +47,4 @@ const Header = () => {
   );
 };
 
-export default Header;
+export default React.memo(Header);
